accessible/tests/mochitest/table/test_headers_table.html

<!DOCTYPE HTML PUBLIC "-//w3c//dtd html 4.0 transitional//en">
<html>
<head>
  <title>Table header information cells for HTML table</title>
  <meta http-equiv="content-type" content="text/html; charset=UTF-8">
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../table.js"></script>

  <script type="application/javascript">

    function doTest()
    {
      //////////////////////////////////////////////////////////////////////////
      // column header from thead and row header from @scope inside of tfoot

      var headerInfoMap = [
        {
          cell: "table1_cell_1",
          rowHeaderCells: [],
          columnHeaderCells: [ "table1_weekday", "table1_date" ]
        },
        {
          cell: "table1_cell_2",
          rowHeaderCells: [],
          columnHeaderCells: [ "table1_day", "table1_date" ]
        },
        {
          cell: "table1_cell_3",
          rowHeaderCells: [],
          columnHeaderCells: [ "table1_qty" ]
        },
        {
          cell: "table1_cell_4",
          rowHeaderCells: [],
          columnHeaderCells: [ "table1_weekday", "table1_date" ]
        },
        {
          cell: "table1_cell_5",
          rowHeaderCells: [],
          columnHeaderCells: [ "table1_day", "table1_date" ]
        },
        {
          cell: "table1_cell_6",
          rowHeaderCells: [],
          columnHeaderCells: [ "table1_qty" ]
        },
        {
          cell: "table1_cell_7",
          rowHeaderCells: [ "table1_total" ],
          columnHeaderCells: [ "table1_qty" ]
        }
      ];

      testHeaderCells(headerInfoMap);

      //////////////////////////////////////////////////////////////////////////
      // column and row headers from thead and @scope

      headerInfoMap = [
        {
          cell: "table2_cell_2",
          rowHeaderCells: [ "table2_rh_1" ],
          columnHeaderCells: [ "table2_ch_2" ]
        },
        {
          cell: "table2_cell_3",
          rowHeaderCells: [ "table2_rh_1" ],
          columnHeaderCells: [ "table2_ch_3" ]
        },
        {
          cell: "table2_cell_5",
          rowHeaderCells: [ "table2_rh_2" ],
          columnHeaderCells: [ "table2_ch_2" ]
        },
        {
          cell: "table2_cell_6",
          rowHeaderCells: [ "table2_rh_2" ],
          columnHeaderCells: [ "table2_ch_3" ]
        },
        {
          cell: "table2_rh_1",
          rowHeaderCells: [],
          columnHeaderCells: [ "table2_ch_1" ]
        },
        {
          cell: "table2_rh_2",
          rowHeaderCells: [],
          columnHeaderCells: [ "table2_ch_1" ]
        }
      ];

      testHeaderCells(headerInfoMap);

      //////////////////////////////////////////////////////////////////////////
      // column headers from @headers

      headerInfoMap = [
        {
          cell: "table3_cell_1",
          rowHeaderCells: [],
          columnHeaderCells: [ "table3_ch_1" ]
        },
        {
          cell: "table3_cell_2",
          rowHeaderCells: [],
          columnHeaderCells: [ "table3_ch_2" ]
        }
      ];

      testHeaderCells(headerInfoMap);

      //////////////////////////////////////////////////////////////////////////
      // table consisted of one column

      headerInfoMap = [
        {
          cell: "table4_cell",
          rowHeaderCells: [],
          columnHeaderCells: [ "table4_ch" ]
        }
      ];

      testHeaderCells(headerInfoMap);

      //////////////////////////////////////////////////////////////////////////
      // table consisted of one row

      headerInfoMap = [
        {
          cell: "table5_cell",
          rowHeaderCells: [ "table5_rh" ],
          columnHeaderCells: [ ]
        }
      ];

      testHeaderCells(headerInfoMap);

      //////////////////////////////////////////////////////////////////////////
      // @headers points to table cells

      headerInfoMap = [
        {
          cell: "table6_cell",
          rowHeaderCells: [ "table6_rh" ],
          columnHeaderCells: [ "table6_ch" ]
        }
      ];

      testHeaderCells(headerInfoMap);

      //////////////////////////////////////////////////////////////////////////
      // @scope="rowgroup" and @scope="row"

      headerInfoMap = [
        {
          cell: "t7_r1c1",
          rowHeaderCells: [ "t7_Females", "t7_Mary" ],
          columnHeaderCells: [ "t7_1km" ]
        },
        {
          cell: "t7_r1c2",
          rowHeaderCells: [ "t7_Females", "t7_Mary" ],
          columnHeaderCells: [ "t7_5km" ]
        },
        {
          cell: "t7_r1c3",
          rowHeaderCells: [ "t7_Females", "t7_Mary" ],
          columnHeaderCells: [ "t7_10km" ]
        },
        {
          cell: "t7_r2c1",
          rowHeaderCells: [ "t7_Females", "t7_Betsy" ],
          columnHeaderCells: [ "t7_1km" ]
        },
        {
          cell: "t7_r2c2",
          rowHeaderCells: [ "t7_Females", "t7_Betsy" ],
          columnHeaderCells: [ "t7_5km" ]
        },
        {
          cell: "t7_r2c3",
          rowHeaderCells: [ "t7_Females", "t7_Betsy" ],
          columnHeaderCells: [ "t7_10km" ]
        },
        {
          cell: "t7_r3c1",
          rowHeaderCells: [ "t7_Males", "t7_Matt" ],
          columnHeaderCells: [ "t7_1km" ]
        },
        {
          cell: "t7_r3c2",
          rowHeaderCells: [ "t7_Males", "t7_Matt" ],
          columnHeaderCells: [ "t7_5km" ]
        },
        {
          cell: "t7_r3c3",
          rowHeaderCells: [ "t7_Males", "t7_Matt" ],
          columnHeaderCells: [ "t7_10km" ]
        },
        {
          cell: "t7_r4c1",
          rowHeaderCells: [ "t7_Males", "t7_Todd" ],
          columnHeaderCells: [ "t7_1km" ]
        },
        {
          cell: "t7_r4c2",
          rowHeaderCells: [ "t7_Males", "t7_Todd" ],
          columnHeaderCells: [ "t7_5km" ]
        },
        {
          cell: "t7_r4c3",
          rowHeaderCells: [ "t7_Males", "t7_Todd" ],
          columnHeaderCells: [ "t7_10km" ]
        }
      ];

      testHeaderCells(headerInfoMap);

      //////////////////////////////////////////////////////////////////////////
      // @scope="colgroup" and @scope="col"

      headerInfoMap = [
        {
          cell: "t8_r1c1",
          rowHeaderCells: [ "t8_1km" ],
          columnHeaderCells: [ "t7_Females", "t7_Mary" ]
        },
        {
          cell: "t8_r1c2",
          rowHeaderCells: [ "t8_5km" ],
          columnHeaderCells: [ "t8_Females", "t8_Mary" ]
        },
        {
          cell: "t8_r1c3",
          rowHeaderCells: [ "t8_10km" ],
          columnHeaderCells: [ "t8_Females", "t8_Mary" ]
        },
        {
          cell: "t8_r1c4",
          rowHeaderCells: [ "t8_1km" ],
          columnHeaderCells: [ "t8_Females", "t8_Betsy" ]
        },
        {
          cell: "t8_r2c1",
          rowHeaderCells: [ "t8_5km" ],
          columnHeaderCells: [ "t8_Females", "t8_Betsy" ]
        },
        {
          cell: "t8_r2c2",
          rowHeaderCells: [ "t8_10km" ],
          columnHeaderCells: [ "t8_Females", "t8_Betsy" ]
        },
        {
          cell: "t8_r2c3",
          rowHeaderCells: [ "t8_1km" ],
          columnHeaderCells: [ "t8_Males", "t8_Matt" ]
        },
        {
          cell: "t8_r2c4",
          rowHeaderCells: [ "t8_5km" ],
          columnHeaderCells: [ "t8_Males", "t8_Matt" ]
        },
        {
          cell: "t8_r3c1",
          rowHeaderCells: [ "t8_10km" ],
          columnHeaderCells: [ "t8_Males", "t8_Matt" ]
        },
        {
          cell: "t8_r3c2",
          rowHeaderCells: [ "t8_1km" ],
          columnHeaderCells: [ "t8_Males", "t8_Todd" ]
        },
        {
          cell: "t8_r3c3",
          rowHeaderCells: [ "t8_5km" ],
          columnHeaderCells: [ "t8_Males", "t8_Todd" ]
        },
        {
          cell: "t8_r3c4",
          rowHeaderCells: [ "t8_10km" ],
          columnHeaderCells: [ "t8_Males", "t8_Todd" ]
        }
      ];

      testHeaderCells(headerInfoMap);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>

<body>
  <a target="_blank"
     title="implement IAccessibleTable2"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=512424">
    Bug 512424
  </a>
  <a target="_blank"
     title="Table headers not associated when header is a td element with no scope"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=704465">
    Bug 704465
  </a>
  <a target="_blank"
     title="Support rowgroup and colgroup HTML scope"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=1141978">
    Bug 1141978
  </a>

  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <table id="table1" border="1">
    <thead>
      <tr>
        <th id="table1_date" colspan="2">Date</th>
        <th id="table1_qty" rowspan="2">Qty</th>
      </tr>
      <tr>
        <th id="table1_weekday">Weekday</th>
        <th id="table1_day">Day</th>
      </tr>
    </thead>
    <tbody>
      <tr>
        <td id="table1_cell_1">Mon</td>
        <td id="table1_cell_2">1</td>
        <td id="table1_cell_3">20</td>
      </tr>
      <tr>
        <td id="table1_cell_4">Thu</td>
        <td id="table1_cell_5">2</td>
        <td id="table1_cell_6">15</td>
      </tr>
    </tbody>
    <tfoot>
      <tr>
        <th id="table1_total" scope="row" colspan="2">Total</th>
        <td id="table1_cell_7">35</td>
      </tr>
    </tfoot>
  </table>

  <table id="table2" border="1">
    <thead>
      <tr>
        <th id="table2_ch_1">col1</th>
        <th id="table2_ch_2">col2</th>
        <td id="table2_ch_3" scope="col">col3</td>
      </tr>
    </thead>
    <tbody>
      <tr>
        <th id="table2_rh_1">row1</th>
        <td id="table2_cell_2">cell1</td>
        <td id="table2_cell_3">cell2</td>
      </tr>
      <tr>
        <td id="table2_rh_2" scope="row">row2</td>
        <td id="table2_cell_5">cell3</td>
        <td id="table2_cell_6">cell4</td>
      </tr>
    </tbody>
  </table>

  <table id="table3" border="1">
    <tr>
      <td id="table3_cell_1" headers="table3_ch_1">cell1</td>
      <td id="table3_cell_2" headers="table3_ch_2">cell2</td>
    </tr>
    <tr>
      <td id="table3_ch_1" scope="col">col1</td>
      <td id="table3_ch_2" scope="col">col2</td>
    </tr>
  </table>

  <table id="table4">
    <thead>
      <tr>
        <th id="table4_ch">colheader</th>
      </tr>
    </thead>
    <tbody>
      <tr>
        <td id="table4_cell">bla</td>
      </tr>
    </tbody>
  </table>

  <table id="table5">
    <tr>
      <th id="table5_rh">rowheader</th>
      <td id="table5_cell">cell</td>
    </tr>
  </table>

  <table id="table6">
    <tr>
      <td>empty cell</th>
      <td id="table6_ch">colheader</td>
    </tr>
    <tr>
      <td id="table6_rh">rowheader</th>
      <td id="table6_cell" headers="table6_ch table6_rh">cell</td>
    </tr>
  </table>

  <table id="table7" class="data complex" border="1">
    <caption>Version 1 with rowgroup</caption>
    <thead>
    <tr>
      <td colspan="2">&nbsp;</td>
      <th id="t7_1km" scope="col">1 km</th>
      <th id="t7_5km" scope="col">5 km</th>
      <th id="t7_10km" scope="col">10 km</th>
    </tr>
    </thead>
    <tbody>
      <tr>
        <th id="t7_Females" rowspan="2" scope="rowgroup">Females</th>
        <th id="t7_Mary" scope="row">Mary</th>
        <td id="t7_r1c1">8:32</td>
        <td id="t7_r1c2">28:04</td>
        <td id="t7_r1c3">1:01:16</td>
      </tr>
      <tr>
        <th id="t7_Betsy" scope="row">Betsy</th>
        <td id="t7_r2c1">7:43</td>
        <td id="t7_r2c2">26:47</td>
        <td id="t7_r2c3">55:38</td>
      </tr>
      <tr>
        <th id="t7_Males" rowspan="2" scope="rowgroup">Males</th>
        <th id="t7_Matt" scope="row">Matt</th>
        <td id="t7_r3c1">7:55</td>
        <td id="t7_r3c2">27:29</td>
        <td id="t7_r3c3">57:04</td>
      </tr>
      <tr>
        <th id="t7_Todd" scope="row">Todd</th>
        <td id="t7_r4c1">7:01</td>
        <td id="t7_r4c2">24:21</td>
        <td id="t7_r4c3">50:35</td>
      </tr>
    </tbody>
  </table>

  <table id="table8" class="data complex" border="1">
    <caption>Version 2 with colgroup</caption>
    <thead>
      <tr>
        <td rowspan="2">&nbsp;</td>
        <th id="t8_Females" colspan="2" scope="colgroup">Females</th>
        <th id="t8_Males" colspan="2" scope="colgroup">Males</th>
      </tr>
      <tr>
        <th id="t8_Mary" scope="col">Mary</th>
        <th id="t8_Betsy" scope="col">Betsy</th>
        <th id="t8_Matt" scope="col">Matt</th>
        <th id="t8_Todd" scope="col">Todd</th>
      </tr>
    </thead>
    <tbody>
      <tr>
        <th id="t8_1km" scope="row">1 km</th>
        <td id="t8_r1c1">8:32</td>
        <td id="t8_r1c2">7:43</td>
        <td id="t8_r1c3">7:55</td>
        <td id="t8_r1c4">7:01</td>
      </tr>
      <tr>
        <th id="t8_5km" scope="row">5 km</th>
        <td id="t8_r2c1">28:04</td>
        <td id="t8_r2c2">26:47</td>
        <td id="t8_r2c3">27:27</td>
        <td id="t8_r2c4">24:21</td>
      </tr>
      <tr>
        <th id="t8_10km" scope="row">10 km</th>
        <td id="t8_r3c1">1:01:16</td>
        <td id="t8_r3c2">55:38</td>
        <td id="t8_r3c3">57:04</td>
        <td id="t8_r3c4">50:35</td>
      </tr>

    </tbody>
  </table>

</body>
</html>
